refactor: nest toolbar group inside toolbar builder
